pyproject.toml
[tool.poetry]
name = "wellness"
version = "0.1.0"
description = ""
authors = ["haeunlee <gkdmsgkdms22@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.113.0"
uvicorn = "^0.30.6"
requests = "^2.32.3"
pillow = "^10.4.0"
sqlalchemy = "^2.0.34"
aiofiles = "^24.1.0"
torch = "^2.4.1"
torchvision = "^0.19.1"
boto3 = "^1.35.13"
python-multipart = "^0.0.9"
timm = "^1.0.9"
httpx = "^0.27.2"
pytest-asyncio = "^0.24.0"
python-dotenv = "^1.0.1"
alembic = "^1.13.2"
psycopg2-binary = "2.9.9"
urllib3 = ">=1.26.0, <1.27"
pydantic = {extras = ["email"], version = "^2.9.1"}
pyjwt = "^2.9.0"
passlib = "^1.7.4"
python-jose = "^3.3.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
